requirements.txt
django==4.2.7
djangorestframework==3.14.0
python-dotenv==1.0.0
requests==2.31.0
django-cors-headers==4.3.0
gTTS==2.3.2
